fix(check): key in-group types in any order; prune text-free composites
Review findings on PR #384 (QSL-156 A4b):

- A type rebuilt outside its recursion group after the group settled is
  now its member: settle maps each member's out-of-group body to the
  member key and insert consults that map. Option<Node> is G17 whichever
  function names it first, and Sequence<Tree>[0,5] is over G8's base.
  The declaration-order-dependent settled-type list is removed.
- The FR-093 text-leaf walk does not enter a composite from which no Text
  is reachable; reachability is memoized per declaration, faults
  UnknownComposite on an unknown composite, and every composite entry
  charges the work meter. FR-093 rule 4 states the pruning.
- A recursion leaf holds its depth as usize, so no usize-to-u64 conversion
  can fail and be reported as InvalidGroup.
- Tests: eqo alone and eqo before eq (P15/P16), function-order
  permutations, Sequence<Tree>[0,5] over G8, a twelve-record text-free
  cluster, contains over List with no leaves.
- qsl-bench recursion fixtures use PackageDeclarations::new(owner()).

+++ qsl-semantics/src/check/node_key/mod.rs
@@ -625,7 +625,7 @@ pub enum LeafSegment {
     Inner,
     /// `recursion:<d>`: the path reenters the composite it entered after
     /// `d` segments.
-    Recursion(u64),
+    Recursion(usize),
 }
 
 impl Serialize for LeafSegment {
